day3/lab3-2_monitoring-cicd/requirements.txt
# Lab 3-2: Monitoring & CI/CD Pipeline
# Python dependencies

# Prometheus & Monitoring
prometheus-client>=0.17.0
requests>=2.28.0
urllib3>=1.26.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0
joblib>=1.2.0

# API (optional)
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# MLflow (optional)
mlflow>=2.9.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# YAML Processing
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Kubernetes (for scripts)
kubernetes>=28.0.0
